+++ forms/DifficultyScreen.Designer.cs
@@ -104,7 +104,7 @@
             this.label1.Font = new System.Drawing.Font("Segoe UI", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.label1.Location = new System.Drawing.Point(10, 17);
             this.label1.Name = "label1";
-            this.label1.Size = new System.Drawing.Size(210, 21);
+            this.label1.Size = new System.Drawing.Size(262, 28);
             this.label1.TabIndex = 59;
             this.label1.Text = "Custom Difficulty Settings";
             // 
@@ -112,7 +112,7 @@
             // 
             this.txtDifficultyName.Location = new System.Drawing.Point(103, 48);
             this.txtDifficultyName.Name = "txtDifficultyName";
-            this.txtDifficultyName.Size = new System.Drawing.Size(135, 22);
+            this.txtDifficultyName.Size = new System.Drawing.Size(135, 26);
             this.txtDifficultyName.TabIndex = 2;
             // 
             // label5
@@ -120,143 +120,143 @@
             this.label5.AutoSize = true;
             this.label5.Location = new System.Drawing.Point(12, 51);
             this.label5.Name = "label5";
-            this.label5.Size = new System.Drawing.Size(85, 13);
+            this.label5.Size = new System.Drawing.Size(102, 19);
             this.label5.TabIndex = 81;
             this.label5.Text = "Difficulty Name";
             // 
             // label2
             // 
             this.label2.AutoSize = true;
-            this.label2.Location = new System.Drawing.Point(277, 148);
+            this.label2.Location = new System.Drawing.Point(277, 162);
             this.label2.Name = "label2";
-            this.label2.Size = new System.Drawing.Size(95, 13);
+            this.label2.Size = new System.Drawing.Size(111, 19);
             this.label2.TabIndex = 83;
             this.label2.Text = "Spawn Multiplier";
             // 
             // txtEnemyHP
             // 
-            this.txtEnemyHP.Location = new System.Drawing.Point(376, 116);
+            this.txtEnemyHP.Location = new System.Drawing.Point(376, 130);
             this.txtEnemyHP.Name = "txtEnemyHP";
-            this.txtEnemyHP.Size = new System.Drawing.Size(29, 22);
+            this.txtEnemyHP.Size = new System.Drawing.Size(29, 26);
             this.txtEnemyHP.TabIndex = 15;
             this.txtEnemyHP.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
             // label3
             // 
             this.label3.AutoSize = true;
-            this.label3.Location = new System.Drawing.Point(277, 119);
+            this.label3.Location = new System.Drawing.Point(277, 133);
             this.label3.Name = "label3";
-            this.label3.Size = new System.Drawing.Size(21, 13);
+            this.label3.Size = new System.Drawing.Size(27, 19);
             this.label3.TabIndex = 85;
             this.label3.Text = "HP";
             // 
             // txtEnemySpawn
             // 
-            this.txtEnemySpawn.Location = new System.Drawing.Point(376, 145);
+            this.txtEnemySpawn.Location = new System.Drawing.Point(376, 159);
             this.txtEnemySpawn.Name = "txtEnemySpawn";
-            this.txtEnemySpawn.Size = new System.Drawing.Size(29, 22);
+            this.txtEnemySpawn.Size = new System.Drawing.Size(29, 26);
             this.txtEnemySpawn.TabIndex = 16;
             this.txtEnemySpawn.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
             // label4
             // 
             this.label4.AutoSize = true;
-            this.label4.Location = new System.Drawing.Point(277, 177);
+            this.label4.Location = new System.Drawing.Point(277, 191);
             this.label4.Name = "label4";
-            this.label4.Size = new System.Drawing.Size(52, 13);
+            this.label4.Size = new System.Drawing.Size(64, 19);
             this.label4.TabIndex = 87;
             this.label4.Text = "AFK Rate";
             // 
             // txtEnemyAFK
             // 
-            this.txtEnemyAFK.Location = new System.Drawing.Point(376, 174);
+            this.txtEnemyAFK.Location = new System.Drawing.Point(376, 188);
             this.txtEnemyAFK.Name = "txtEnemyAFK";
-            this.txtEnemyAFK.Size = new System.Drawing.Size(29, 22);
+            this.txtEnemyAFK.Size = new System.Drawing.Size(29, 26);
             this.txtEnemyAFK.TabIndex = 17;
             this.txtEnemyAFK.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
             // label6
             // 
             this.label6.AutoSize = true;
-            this.label6.Location = new System.Drawing.Point(277, 206);
+            this.label6.Location = new System.Drawing.Point(277, 220);
             this.label6.Name = "label6";
-            this.label6.Size = new System.Drawing.Size(66, 13);
+            this.label6.Size = new System.Drawing.Size(81, 19);
             this.label6.TabIndex = 89;
             this.label6.Text = "Aggro Limit";
             // 
             // txtEnemyAggro
             // 
-            this.txtEnemyAggro.Location = new System.Drawing.Point(376, 203);
+            this.txtEnemyAggro.Location = new System.Drawing.Point(376, 217);
             this.txtEnemyAggro.Name = "txtEnemyAggro";
-            this.txtEnemyAggro.Size = new System.Drawing.Size(29, 22);
+            this.txtEnemyAggro.Size = new System.Drawing.Size(29, 26);
             this.txtEnemyAggro.TabIndex = 18;
             this.txtEnemyAggro.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
             // label7
             // 
             this.label7.AutoSize = true;
-            this.label7.Location = new System.Drawing.Point(12, 206);
+            this.label7.Location = new System.Drawing.Point(12, 220);
             this.label7.Name = "label7";
-            this.label7.Size = new System.Drawing.Size(75, 13);
+            this.label7.Size = new System.Drawing.Size(93, 19);
             this.label7.TabIndex = 97;
             this.label7.Text = "Lives (Arcade)";
             // 
             // txtPlayerLivesArcade
             // 
-            this.txtPlayerLivesArcade.Location = new System.Drawing.Point(110, 203);
+            this.txtPlayerLivesArcade.Location = new System.Drawing.Point(110, 217);
             this.txtPlayerLivesArcade.Name = "txtPlayerLivesArcade";
-            this.txtPlayerLivesArcade.Size = new System.Drawing.Size(29, 22);
+            this.txtPlayerLivesArcade.Size = new System.Drawing.Size(29, 26);
             this.txtPlayerLivesArcade.TabIndex = 6;
             this.txtPlayerLivesArcade.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
             // label8
             // 
             this.label8.AutoSize = true;
-            this.label8.Location = new System.Drawing.Point(12, 177);
+            this.label8.Location = new System.Drawing.Point(12, 191);
             this.label8.Name = "label8";
-            this.label8.Size = new System.Drawing.Size(31, 13);
+            this.label8.Size = new System.Drawing.Size(39, 19);
             this.label8.TabIndex = 95;
             this.label8.Text = "Lives";
             // 
             // txtPlayerLives
             // 
-            this.txtPlayerLives.Location = new System.Drawing.Point(110, 174);
+            this.txtPlayerLives.Location = new System.Drawing.Point(110, 188);
             this.txtPlayerLives.Name = "txtPlayerLives";
-            this.txtPlayerLives.Size = new System.Drawing.Size(29, 22);
+            this.txtPlayerLives.Size = new System.Drawing.Size(29, 26);
             this.txtPlayerLives.TabIndex = 5;
             this.txtPlayerLives.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
             // label9
             // 
             this.label9.AutoSize = true;
-            this.label9.Location = new System.Drawing.Point(12, 119);
+            this.label9.Location = new System.Drawing.Point(12, 133);
             this.label9.Name = "label9";
-            this.label9.Size = new System.Drawing.Size(49, 13);
+            this.label9.Size = new System.Drawing.Size(58, 19);
             this.label9.TabIndex = 93;
             this.label9.Text = "Defense";
             // 
             // txtPlayerGreen
             // 
-            this.txtPlayerGreen.Location = new System.Drawing.Point(110, 145);
+            this.txtPlayerGreen.Location = new System.Drawing.Point(110, 159);
             this.txtPlayerGreen.Name = "txtPlayerGreen";
-            this.txtPlayerGreen.Size = new System.Drawing.Size(29, 22);
+            this.txtPlayerGreen.Size = new System.Drawing.Size(29, 26);
             this.txtPlayerGreen.TabIndex = 4;
             this.txtPlayerGreen.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
             // label10
             // 
             this.label10.AutoSize = true;
-            this.label10.Location = new System.Drawing.Point(12, 148);
+            this.label10.Location = new System.Drawing.Point(12, 162);
             this.label10.Name = "label10";
-            this.label10.Size = new System.Drawing.Size(97, 13);
+            this.label10.Size = new System.Drawing.Size(117, 19);
             this.label10.TabIndex = 91;
             this.label10.Text = "Green Health Use";
             // 
             // txtPlayerDefense
             // 
-            this.txtPlayerDefense.Location = new System.Drawing.Point(110, 116);
+            this.txtPlayerDefense.Location = new System.Drawing.Point(110, 130);
             this.txtPlayerDefense.Name = "txtPlayerDefense";
-            this.txtPlayerDefense.Size = new System.Drawing.Size(29, 22);
+            this.txtPlayerDefense.Size = new System.Drawing.Size(29, 26);
             this.txtPlayerDefense.TabIndex = 3;
             this.txtPlayerDefense.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
@@ -264,9 +264,9 @@
             // 
             this.label11.AutoSize = true;
             this.label11.Font = new System.Drawing.Font("Segoe UI", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.label11.Location = new System.Drawing.Point(10, 90);
+            this.label11.Location = new System.Drawing.Point(10, 104);
             this.label11.Name = "label11";
-            this.label11.Size = new System.Drawing.Size(58, 21);
+            this.label11.Size = new System.Drawing.Size(71, 28);
             this.label11.TabIndex = 98;
             this.label11.Text = "Player";
             // 
@@ -274,26 +274,26 @@
             // 
             this.label12.AutoSize = true;
             this.label12.Font = new System.Drawing.Font("Segoe UI", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.label12.Location = new System.Drawing.Point(276, 90);
+            this.label12.Location = new System.Drawing.Point(276, 104);
             this.label12.Name = "label12";
-            this.label12.Size = new System.Drawing.Size(74, 21);
+            this.label12.Size = new System.Drawing.Size(90, 28);
             this.label12.TabIndex = 99;
             this.label12.Text = "Enemies";
             // 
             // label13
             // 
             this.label13.AutoSize = true;
-            this.label13.Location = new System.Drawing.Point(12, 235);
+            this.label13.Location = new System.Drawing.Point(12, 249);
             this.label13.Name = "label13";
-            this.label13.Size = new System.Drawing.Size(32, 13);
+            this.label13.Size = new System.Drawing.Size(39, 19);
             this.label13.TabIndex = 101;
             this.label13.Text = "Stars";
             // 
             // txtPlayerStars
             // 
-            this.txtPlayerStars.Location = new System.Drawing.Point(110, 232);
+            this.txtPlayerStars.Location = new System.Drawing.Point(110, 246);
             this.txtPlayerStars.Name = "txtPlayerStars";
-            this.txtPlayerStars.Size = new System.Drawing.Size(29, 22);
+            this.txtPlayerStars.Size = new System.Drawing.Size(29, 26);
             this.txtPlayerStars.TabIndex = 7;
             this.txtPlayerStars.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
@@ -303,75 +303,75 @@
             this.cmbDifficultyCollection.FormattingEnabled = true;
             this.cmbDifficultyCollection.Location = new System.Drawing.Point(244, 48);
             this.cmbDifficultyCollection.Name = "cmbDifficultyCollection";
-            this.cmbDifficultyCollection.Size = new System.Drawing.Size(121, 21);
+            this.cmbDifficultyCollection.Size = new System.Drawing.Size(121, 27);
             this.cmbDifficultyCollection.TabIndex = 106;
             this.cmbDifficultyCollection.SelectedIndexChanged += new System.EventHandler(this.cmbDifficultyCollection_SelectedIndexChanged);
             // 
             // label21
             // 
             this.label21.AutoSize = true;
-            this.label21.Location = new System.Drawing.Point(12, 322);
+            this.label21.Location = new System.Drawing.Point(12, 336);
             this.label21.Name = "label21";
-            this.label21.Size = new System.Drawing.Size(45, 13);
+            this.label21.Size = new System.Drawing.Size(54, 19);
             this.label21.TabIndex = 116;
             this.label21.Text = "Hitstun";
             // 
             // txtPlayerHitstun
             // 
-            this.txtPlayerHitstun.Location = new System.Drawing.Point(110, 319);
+            this.txtPlayerHitstun.Location = new System.Drawing.Point(110, 333);
             this.txtPlayerHitstun.Name = "txtPlayerHitstun";
-            this.txtPlayerHitstun.Size = new System.Drawing.Size(29, 22);
+            this.txtPlayerHitstun.Size = new System.Drawing.Size(29, 26);
             this.txtPlayerHitstun.TabIndex = 13;
             this.txtPlayerHitstun.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
             // label22
             // 
             this.label22.AutoSize = true;
-            this.label22.Location = new System.Drawing.Point(12, 293);
+            this.label22.Location = new System.Drawing.Point(12, 307);
             this.label22.Name = "label22";
-            this.label22.Size = new System.Drawing.Size(45, 13);
+            this.label22.Size = new System.Drawing.Size(54, 19);
             this.label22.TabIndex = 114;
             this.label22.Text = "Hitstop";
             // 
             // txtPlayerHitstop
             // 
-            this.txtPlayerHitstop.Location = new System.Drawing.Point(110, 290);
+            this.txtPlayerHitstop.Location = new System.Drawing.Point(110, 304);
             this.txtPlayerHitstop.Name = "txtPlayerHitstop";
-            this.txtPlayerHitstop.Size = new System.Drawing.Size(29, 22);
+            this.txtPlayerHitstop.Size = new System.Drawing.Size(29, 26);
             this.txtPlayerHitstop.TabIndex = 10;
             this.txtPlayerHitstop.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
             // label23
             // 
             this.label23.AutoSize = true;
-            this.label23.Location = new System.Drawing.Point(277, 322);
+            this.label23.Location = new System.Drawing.Point(277, 336);
             this.label23.Name = "label23";
-            this.label23.Size = new System.Drawing.Size(45, 13);
+            this.label23.Size = new System.Drawing.Size(54, 19);
             this.label23.TabIndex = 120;
             this.label23.Text = "Hitstun";
             // 
             // txtEnemyHitstun
             // 
-            this.txtEnemyHitstun.Location = new System.Drawing.Point(376, 319);
+            this.txtEnemyHitstun.Location = new System.Drawing.Point(376, 333);
             this.txtEnemyHitstun.Name = "txtEnemyHitstun";
-            this.txtEnemyHitstun.Size = new System.Drawing.Size(29, 22);
+            this.txtEnemyHitstun.Size = new System.Drawing.Size(29, 26);
             this.txtEnemyHitstun.TabIndex = 25;
             this.txtEnemyHitstun.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
             // label24
             // 
             this.label24.AutoSize = true;
-            this.label24.Location = new System.Drawing.Point(277, 293);
+            this.label24.Location = new System.Drawing.Point(277, 307);
             this.label24.Name = "label24";
-            this.label24.Size = new System.Drawing.Size(45, 13);
+            this.label24.Size = new System.Drawing.Size(54, 19);
             this.label24.TabIndex = 118;
             this.label24.Text = "Hitstop";
             // 
             // txtEnemyHitstop
             // 
-            this.txtEnemyHitstop.Location = new System.Drawing.Point(376, 290);
+            this.txtEnemyHitstop.Location = new System.Drawing.Point(376, 304);
             this.txtEnemyHitstop.Name = "txtEnemyHitstop";
-            this.txtEnemyHitstop.Size = new System.Drawing.Size(29, 22);
+            this.txtEnemyHitstop.Size = new System.Drawing.Size(29, 26);
             this.txtEnemyHitstop.TabIndex = 22;
             this.txtEnemyHitstop.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
@@ -379,9 +379,9 @@
             // 
             this.textBox1.BackColor = System.Drawing.SystemColors.Window;
             this.textBox1.Enabled = false;
-            this.textBox1.Location = new System.Drawing.Point(138, 116);
+            this.textBox1.Location = new System.Drawing.Point(138, 130);
             this.textBox1.Name = "textBox1";
-            this.textBox1.Size = new System.Drawing.Size(19, 22);
+            this.textBox1.Size = new System.Drawing.Size(19, 26);
             this.textBox1.TabIndex = 122;
             this.textBox1.Text = "%";
             // 
@@ -389,9 +389,9 @@
             // 
             this.textBox2.BackColor = System.Drawing.SystemColors.Window;
             this.textBox2.Enabled = false;
-            this.textBox2.Location = new System.Drawing.Point(138, 145);
+            this.textBox2.Location = new System.Drawing.Point(138, 159);
             this.textBox2.Name = "textBox2";
-            this.textBox2.Size = new System.Drawing.Size(19, 22);
+            this.textBox2.Size = new System.Drawing.Size(19, 26);
             this.textBox2.TabIndex = 123;
             this.textBox2.Text = "%";
             // 
@@ -399,9 +399,9 @@
             // 
             this.textBox4.BackColor = System.Drawing.SystemColors.Window;
             this.textBox4.Enabled = false;
-            this.textBox4.Location = new System.Drawing.Point(404, 174);
+            this.textBox4.Location = new System.Drawing.Point(404, 188);
             this.textBox4.Name = "textBox4";
-            this.textBox4.Size = new System.Drawing.Size(19, 22);
+            this.textBox4.Size = new System.Drawing.Size(19, 26);
             this.textBox4.TabIndex = 124;
             this.textBox4.Text = "%";
             // 
@@ -413,9 +413,9 @@
             "Default",
             "Frames",
             "%"});
-            this.cmbPlayerHitstopValuesOptions.Location = new System.Drawing.Point(138, 291);
+            this.cmbPlayerHitstopValuesOptions.Location = new System.Drawing.Point(138, 305);
             this.cmbPlayerHitstopValuesOptions.Name = "cmbPlayerHitstopValuesOptions";
-            this.cmbPlayerHitstopValuesOptions.Size = new System.Drawing.Size(107, 21);
+            this.cmbPlayerHitstopValuesOptions.Size = new System.Drawing.Size(107, 27);
             this.cmbPlayerHitstopValuesOptions.TabIndex = 9;
             this.cmbPlayerHitstopValuesOptions.SelectedIndexChanged += new System.EventHandler(this.cmbPlayerHitstopValuesOptions_SelectedIndexChanged);
             // 
@@ -427,9 +427,9 @@
             "Default",
             "Frames",
             "%"});
-            this.cmbPlayerHitstunValuesOptions.Location = new System.Drawing.Point(138, 320);
+            this.cmbPlayerHitstunValuesOptions.Location = new System.Drawing.Point(138, 334);
             this.cmbPlayerHitstunValuesOptions.Name = "cmbPlayerHitstunValuesOptions";
-            this.cmbPlayerHitstunValuesOptions.Size = new System.Drawing.Size(107, 21);
+            this.cmbPlayerHitstunValuesOptions.Size = new System.Drawing.Size(107, 27);
             this.cmbPlayerHitstunValuesOptions.TabIndex = 12;
             this.cmbPlayerHitstunValuesOptions.SelectedIndexChanged += new System.EventHandler(this.cmbPlayerHitstunValuesOptions_SelectedIndexChanged);
             // 
@@ -441,9 +441,9 @@
             "Default",
             "Frames",
             "%"});
-            this.cmbEnemyHitstunValuesOptions.Location = new System.Drawing.Point(404, 320);
+            this.cmbEnemyHitstunValuesOptions.Location = new System.Drawing.Point(404, 334);
             this.cmbEnemyHitstunValuesOptions.Name = "cmbEnemyHitstunValuesOptions";
-            this.cmbEnemyHitstunValuesOptions.Size = new System.Drawing.Size(107, 21);
+            this.cmbEnemyHitstunValuesOptions.Size = new System.Drawing.Size(107, 27);
             this.cmbEnemyHitstunValuesOptions.TabIndex = 24;
             this.cmbEnemyHitstunValuesOptions.SelectedIndexChanged += new System.EventHandler(this.cmbEnemyHitstunValuesOptions_SelectedIndexChanged);
             // 
@@ -455,26 +455,26 @@
             "Default",
             "Frames",
             "%"});
-            this.cmbEnemyHitstopValuesOptions.Location = new System.Drawing.Point(404, 291);
+            this.cmbEnemyHitstopValuesOptions.Location = new System.Drawing.Point(404, 305);
             this.cmbEnemyHitstopValuesOptions.Name = "cmbEnemyHitstopValuesOptions";
-            this.cmbEnemyHitstopValuesOptions.Size = new System.Drawing.Size(107, 21);
+            this.cmbEnemyHitstopValuesOptions.Size = new System.Drawing.Size(107, 27);
             this.cmbEnemyHitstopValuesOptions.TabIndex = 21;
             this.cmbEnemyHitstopValuesOptions.SelectedIndexChanged += new System.EventHandler(this.cmbEnemyHitstopValuesOptions_SelectedIndexChanged);
             // 
             // label16
             // 
             this.label16.AutoSize = true;
-            this.label16.Location = new System.Drawing.Point(12, 264);
+            this.label16.Location = new System.Drawing.Point(12, 278);
             this.label16.Name = "label16";
-            this.label16.Size = new System.Drawing.Size(96, 13);
+            this.label16.Size = new System.Drawing.Size(117, 19);
             this.label16.TabIndex = 138;
             this.label16.Text = "Movement Speed";
             // 
             // txtPlayerSpeed
             // 
-            this.txtPlayerSpeed.Location = new System.Drawing.Point(110, 261);
+            this.txtPlayerSpeed.Location = new System.Drawing.Point(110, 275);
             this.txtPlayerSpeed.Name = "txtPlayerSpeed";
-            this.txtPlayerSpeed.Size = new System.Drawing.Size(29, 22);
+            this.txtPlayerSpeed.Size = new System.Drawing.Size(29, 26);
             this.txtPlayerSpeed.TabIndex = 8;
             this.txtPlayerSpeed.TextChanged += new System.EventHandler(this.ProcessInput);
             this.txtPlayerSpeed.MouseHover += new System.EventHandler(this.txtPlayerSpeed_MouseHover);
@@ -482,17 +482,17 @@
             // label17
             // 
             this.label17.AutoSize = true;
-            this.label17.Location = new System.Drawing.Point(277, 264);
+            this.label17.Location = new System.Drawing.Point(277, 278);
             this.label17.Name = "label17";
-            this.label17.Size = new System.Drawing.Size(96, 13);
+            this.label17.Size = new System.Drawing.Size(117, 19);
             this.label17.TabIndex = 140;
             this.label17.Text = "Movement Speed";
             // 
             // txtEnemySpeed
             // 
-            this.txtEnemySpeed.Location = new System.Drawing.Point(375, 261);
+            this.txtEnemySpeed.Location = new System.Drawing.Point(375, 275);
             this.txtEnemySpeed.Name = "txtEnemySpeed";
-            this.txtEnemySpeed.Size = new System.Drawing.Size(29, 22);
+            this.txtEnemySpeed.Size = new System.Drawing.Size(29, 26);
             this.txtEnemySpeed.TabIndex = 20;
             this.txtEnemySpeed.TextChanged += new System.EventHandler(this.ProcessInput);
             this.txtEnemySpeed.MouseHover += new System.EventHandler(this.txtEnemySpeed_MouseHover);
@@ -501,9 +501,9 @@
             // 
             this.textBox6.BackColor = System.Drawing.SystemColors.Window;
             this.textBox6.Enabled = false;
-            this.textBox6.Location = new System.Drawing.Point(404, 116);
+            this.textBox6.Location = new System.Drawing.Point(404, 130);
             this.textBox6.Name = "textBox6";
-            this.textBox6.Size = new System.Drawing.Size(19, 22);
+            this.textBox6.Size = new System.Drawing.Size(19, 26);
             this.textBox6.TabIndex = 141;
             this.textBox6.Text = "%";
             // 
@@ -511,9 +511,9 @@
             // 
             this.textBox7.BackColor = System.Drawing.SystemColors.Window;
             this.textBox7.Enabled = false;
-            this.textBox7.Location = new System.Drawing.Point(403, 261);
+            this.textBox7.Location = new System.Drawing.Point(403, 275);
             this.textBox7.Name = "textBox7";
-            this.textBox7.Size = new System.Drawing.Size(19, 22);
+            this.textBox7.Size = new System.Drawing.Size(19, 26);
             this.textBox7.TabIndex = 142;
             this.textBox7.Text = "%";
             // 
@@ -521,9 +521,9 @@
             // 
             this.textBox8.BackColor = System.Drawing.SystemColors.Window;
             this.textBox8.Enabled = false;
-            this.textBox8.Location = new System.Drawing.Point(403, 145);
+            this.textBox8.Location = new System.Drawing.Point(403, 159);
             this.textBox8.Name = "textBox8";
-            this.textBox8.Size = new System.Drawing.Size(19, 22);
+            this.textBox8.Size = new System.Drawing.Size(19, 26);
             this.textBox8.TabIndex = 143;
             this.textBox8.Text = "%";
             // 
@@ -531,18 +531,18 @@
             // 
             this.textBox9.BackColor = System.Drawing.SystemColors.Window;
             this.textBox9.Enabled = false;
-            this.textBox9.Location = new System.Drawing.Point(138, 261);
+            this.textBox9.Location = new System.Drawing.Point(138, 275);
             this.textBox9.Name = "textBox9";
-            this.textBox9.Size = new System.Drawing.Size(19, 22);
+            this.textBox9.Size = new System.Drawing.Size(19, 26);
             this.textBox9.TabIndex = 144;
             this.textBox9.Text = "%";
             // 
             // txtPlayerHitstopMin
             // 
             this.txtPlayerHitstopMin.Enabled = false;
-            this.txtPlayerHitstopMin.Location = new System.Drawing.Point(221, 290);
+            this.txtPlayerHitstopMin.Location = new System.Drawing.Point(221, 304);
             this.txtPlayerHitstopMin.Name = "txtPlayerHitstopMin";
-            this.txtPlayerHitstopMin.Size = new System.Drawing.Size(24, 22);
+            this.txtPlayerHitstopMin.Size = new System.Drawing.Size(24, 26);
             this.txtPlayerHitstopMin.TabIndex = 11;
             this.txtPlayerHitstopMin.Visible = false;
             this.txtPlayerHitstopMin.TextChanged += new System.EventHandler(this.ProcessInput);
@@ -550,27 +550,27 @@
             // txtPlayerHitstunMin
             // 
             this.txtPlayerHitstunMin.Enabled = false;
-            this.txtPlayerHitstunMin.Location = new System.Drawing.Point(221, 319);
+            this.txtPlayerHitstunMin.Location = new System.Drawing.Point(221, 333);
             this.txtPlayerHitstunMin.Name = "txtPlayerHitstunMin";
-            this.txtPlayerHitstunMin.Size = new System.Drawing.Size(24, 22);
+            this.txtPlayerHitstunMin.Size = new System.Drawing.Size(24, 26);
             this.txtPlayerHitstunMin.TabIndex = 14;
             this.txtPlayerHitstunMin.Visible = false;
             this.txtPlayerHitstunMin.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
             // txtEnemyHitstopMin
             // 
-            this.txtEnemyHitstopMin.Location = new System.Drawing.Point(487, 290);
+            this.txtEnemyHitstopMin.Location = new System.Drawing.Point(487, 304);
             this.txtEnemyHitstopMin.Name = "txtEnemyHitstopMin";
-            this.txtEnemyHitstopMin.Size = new System.Drawing.Size(24, 22);
+            this.txtEnemyHitstopMin.Size = new System.Drawing.Size(24, 26);
             this.txtEnemyHitstopMin.TabIndex = 23;
             this.txtEnemyHitstopMin.Visible = false;
             this.txtEnemyHitstopMin.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
             // txtEnemyHitstunMin
             // 
-            this.txtEnemyHitstunMin.Location = new System.Drawing.Point(487, 319);
+            this.txtEnemyHitstunMin.Location = new System.Drawing.Point(487, 333);
             this.txtEnemyHitstunMin.Name = "txtEnemyHitstunMin";
-            this.txtEnemyHitstunMin.Size = new System.Drawing.Size(24, 22);
+            this.txtEnemyHitstunMin.Size = new System.Drawing.Size(24, 26);
             this.txtEnemyHitstunMin.TabIndex = 26;
             this.txtEnemyHitstunMin.Visible = false;
             this.txtEnemyHitstunMin.TextChanged += new System.EventHandler(this.ProcessInput);
@@ -580,7 +580,7 @@
             this.label14.AutoSize = true;
             this.label14.Location = new System.Drawing.Point(315, 32);
             this.label14.Name = "label14";
-            this.label14.Size = new System.Drawing.Size(54, 13);
+            this.label14.Size = new System.Drawing.Size(65, 19);
             this.label14.TabIndex = 145;
             this.label14.Text = "Based on";
             // 
@@ -589,7 +589,7 @@
             this.labelPlayerHitstopMinLabel.BackColor = System.Drawing.Color.White;
             this.labelPlayerHitstopMinLabel.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
             this.labelPlayerHitstopMinLabel.Font = new System.Drawing.Font("Segoe UI", 8.25F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.labelPlayerHitstopMinLabel.Location = new System.Drawing.Point(191, 290);
+            this.labelPlayerHitstopMinLabel.Location = new System.Drawing.Point(191, 304);
             this.labelPlayerHitstopMinLabel.Name = "labelPlayerHitstopMinLabel";
             this.labelPlayerHitstopMinLabel.Padding = new System.Windows.Forms.Padding(0, 4, 0, 0);
             this.labelPlayerHitstopMinLabel.Size = new System.Drawing.Size(31, 22);
@@ -602,7 +602,7 @@
             this.labelPlayerHitstunMinLabel.BackColor = System.Drawing.Color.White;
             this.labelPlayerHitstunMinLabel.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
             this.labelPlayerHitstunMinLabel.Font = new System.Drawing.Font("Segoe UI", 8.25F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.labelPlayerHitstunMinLabel.Location = new System.Drawing.Point(191, 319);
+            this.labelPlayerHitstunMinLabel.Location = new System.Drawing.Point(191, 333);
             this.labelPlayerHitstunMinLabel.Name = "labelPlayerHitstunMinLabel";
             this.labelPlayerHitstunMinLabel.Padding = new System.Windows.Forms.Padding(0, 4, 0, 0);
             this.labelPlayerHitstunMinLabel.Size = new System.Drawing.Size(31, 22);
@@ -615,7 +615,7 @@
             this.labelEnemyHitstopMinLabel.BackColor = System.Drawing.Color.White;
             this.labelEnemyHitstopMinLabel.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
             this.labelEnemyHitstopMinLabel.Font = new System.Drawing.Font("Segoe UI", 8.25F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.labelEnemyHitstopMinLabel.Location = new System.Drawing.Point(457, 290);
+            this.labelEnemyHitstopMinLabel.Location = new System.Drawing.Point(457, 304);
             this.labelEnemyHitstopMinLabel.Name = "labelEnemyHitstopMinLabel";
             this.labelEnemyHitstopMinLabel.Padding = new System.Windows.Forms.Padding(0, 4, 0, 0);
             this.labelEnemyHitstopMinLabel.Size = new System.Drawing.Size(31, 22);
@@ -628,7 +628,7 @@
             this.labelEnemyHitstunMinLabel.BackColor = System.Drawing.Color.White;
             this.labelEnemyHitstunMinLabel.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
             this.labelEnemyHitstunMinLabel.Font = new System.Drawing.Font("Segoe UI", 8.25F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.labelEnemyHitstunMinLabel.Location = new System.Drawing.Point(457, 319);
+            this.labelEnemyHitstunMinLabel.Location = new System.Drawing.Point(457, 333);
             this.labelEnemyHitstunMinLabel.Name = "labelEnemyHitstunMinLabel";
             this.labelEnemyHitstunMinLabel.Padding = new System.Windows.Forms.Padding(0, 4, 0, 0);
             this.labelEnemyHitstunMinLabel.Size = new System.Drawing.Size(31, 22);
@@ -639,17 +639,17 @@
             // label15
             // 
             this.label15.AutoSize = true;
-            this.label15.Location = new System.Drawing.Point(149, 206);
+            this.label15.Location = new System.Drawing.Point(149, 220);
             this.label15.Name = "label15";
-            this.label15.Size = new System.Drawing.Size(58, 13);
+            this.label15.Size = new System.Drawing.Size(72, 19);
             this.label15.TabIndex = 156;
             this.label15.Text = "1UP Score";
             // 
             // txtLifeUpArcade
             // 
-            this.txtLifeUpArcade.Location = new System.Drawing.Point(210, 203);
+            this.txtLifeUpArcade.Location = new System.Drawing.Point(210, 217);
             this.txtLifeUpArcade.Name = "txtLifeUpArcade";
-            this.txtLifeUpArcade.Size = new System.Drawing.Size(61, 22);
+            this.txtLifeUpArcade.Size = new System.Drawing.Size(61, 26);
             this.txtLifeUpArcade.TabIndex = 154;
             this.txtLifeUpArcade.TextChanged += new System.EventHandler(this.ProcessInput);
             this.txtLifeUpArcade.MouseHover += new System.EventHandler(this.txtLifeUpArcade_MouseHover);
@@ -657,35 +657,35 @@
             // label18
             // 
             this.label18.AutoSize = true;
-            this.label18.Location = new System.Drawing.Point(149, 177);
+            this.label18.Location = new System.Drawing.Point(149, 191);
             this.label18.Name = "label18";
-            this.label18.Size = new System.Drawing.Size(58, 13);
+            this.label18.Size = new System.Drawing.Size(72, 19);
             this.label18.TabIndex = 155;
             this.label18.Text = "1UP Score";
             // 
             // txtLifeUp
             // 
-            this.txtLifeUp.Location = new System.Drawing.Point(210, 174);
+            this.txtLifeUp.Location = new System.Drawing.Point(210, 188);
             this.txtLifeUp.Name = "txtLifeUp";
-            this.txtLifeUp.Size = new System.Drawing.Size(61, 22);
+            this.txtLifeUp.Size = new System.Drawing.Size(61, 26);
             this.txtLifeUp.TabIndex = 153;
             this.txtLifeUp.TextChanged += new System.EventHandler(this.ProcessInput);
             this.txtLifeUp.MouseHover += new System.EventHandler(this.txtLifeUp_MouseHover);
             // 
             // txtEnemySpeedMultiplier
             // 
-            this.txtEnemySpeedMultiplier.Location = new System.Drawing.Point(375, 231);
+            this.txtEnemySpeedMultiplier.Location = new System.Drawing.Point(375, 245);
             this.txtEnemySpeedMultiplier.Name = "txtEnemySpeedMultiplier";
-            this.txtEnemySpeedMultiplier.Size = new System.Drawing.Size(30, 22);
+            this.txtEnemySpeedMultiplier.Size = new System.Drawing.Size(30, 26);
             this.txtEnemySpeedMultiplier.TabIndex = 19;
             this.txtEnemySpeedMultiplier.TextChanged += new System.EventHandler(this.ProcessInput);
             // 
             // label19
             // 
             this.label19.AutoSize = true;
-            this.label19.Location = new System.Drawing.Point(277, 235);
+            this.label19.Location = new System.Drawing.Point(277, 249);
             this.label19.Name = "label19";
-            this.label19.Size = new System.Drawing.Size(71, 13);
+            this.label19.Size = new System.Drawing.Size(85, 19);
             this.label19.TabIndex = 158;
             this.label19.Text = "Speed Boost";
             // 
@@ -693,9 +693,9 @@
             // 
             this.textBox3.BackColor = System.Drawing.SystemColors.Window;
             this.textBox3.Enabled = false;
-            this.textBox3.Location = new System.Drawing.Point(403, 231);
+            this.textBox3.Location = new System.Drawing.Point(403, 245);
             this.textBox3.Name = "textBox3";
-            this.textBox3.Size = new System.Drawing.Size(19, 22);
+            this.textBox3.Size = new System.Drawing.Size(19, 26);
             this.textBox3.TabIndex = 159;
             this.textBox3.Text = "%";
             // 
@@ -711,7 +711,7 @@
             // 
             // DifficultyScreen
             // 
-            this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
+            this.AutoScaleDimensions = new System.Drawing.SizeF(8F, 19F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
             this.ClientSize = new System.Drawing.Size(520, 423);
             this.Controls.Add(this.btnReset);